Renaming library from slug to slugs
This change renames the library from slug (or SLUG) to slugs. This
is done to simplify any future upload to PyPI, which already has a
library named slug registered. The name slugs still fits since the
description is: a Simple, Lightweight User Group Service.
