+++ slugs/plugins.py
