+++ tox.ini
@@ -6,11 +6,11 @@ deps = -r{toxinidir}/requirements.txt
        -r{toxinidir}/test-requirements.txt
 
 commands =
-    coverage run --source=slug/ --omit=slug/tests/* -m pytest --strict slug/tests/unit
+    coverage run --source=slugs/ --omit=slugs/tests/* -m pytest --strict slugs/tests/unit
     coverage report -m
 
 [testenv:pep8]
-commands = flake8 slug/
+commands = flake8 slugs/
 
 [testenv:bandit]
-commands = bandit -r slug -n5
+commands = bandit -r slugs -n5
